backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
stripe==7.11.0
boto3==1.34.22
openai==1.10.0
pillow==10.2.0
requests==2.31.0
asyncpg==0.29.0
aiofiles==23.2.1
python-dotenv==1.0.0
pyotp==2.9.0
qrcode==7.4.2
